Move api unit tests to api module
