+++ api/pom.xml
@@ -37,6 +37,12 @@
             <artifactId>jts-core</artifactId>
             <version>1.14.0</version>
         </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>4.12</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
