+++ api/src/main/java/com/graphhopper/util/NumHelper.java
@@ -39,11 +39,4 @@ public class NumHelper {
         return Double.compare(d1, d2);
     }
 
-    public static boolean equalsEps(float d1, float d2) {
-        return equalsEps(d1, d2, DEFAULT_PRECISION);
-    }
-
-    public static boolean equalsEps(float d1, float d2, float epsilon) {
-        return Math.abs(d1 - d2) < epsilon;
-    }
 }
