+++ api/src/test/java/com/graphhopper/util/AngleCalcTest.java
