+++ api/src/test/java/com/graphhopper/util/FinishInstructionTest.java
