+++ api/src/test/java/com/graphhopper/util/PointListTest.java
