+++ api/src/test/java/com/graphhopper/util/shapes/BBoxTest.java
