+++ api/src/test/java/com/graphhopper/util/shapes/CircleTest.java
