+++ api/src/test/java/com/graphhopper/util/shapes/GHPointTest.java
